requirements.txt
pyTelegramBotAPI
shutilwhich
keyboard
psutil
Pillow
pyautogui
